Add UserGateway
