+++ lib/allegro/webapi.rb
@@ -2,7 +2,7 @@ require 'savon'
 require "allegro/webapi/version"
 require 'allegro/webapi/auction'
 require 'allegro/webapi/client'
-require 'allegro/webapi/user'
+require 'allegro/webapi/user_gateway'
 require 'allegro/webapi/search'
 require 'allegro/webapi/journal_gateway'
 require 'yaml'
